+++ monarchy-example/src/main/java/com/zhuinden/monarchyexample/features/frozen/FrozenFragment.java
@@ -1,0 +1,76 @@
+package com.zhuinden.monarchyexample.features.frozen;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.zhuinden.monarchy.Monarchy;
+import com.zhuinden.monarchyexample.R;
+import com.zhuinden.monarchyexample.RealmDog;
+import com.zhuinden.monarchyexample.application.CustomApplication;
+import com.zhuinden.monarchyexample.utils.BaseFragment;
+
+import java.util.List;
+
+import javax.inject.Inject;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.Observer;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
+/**
+ * Created by Zhuinden on 2020.06.25.
+ */
+
+public class FrozenFragment
+        extends BaseFragment {
+    @BindView(R.id.recycler_view)
+    RecyclerView recyclerView;
+
+    @Inject
+    Monarchy monarchy;
+
+    FrozenDogAdapter frozenDogAdapter;
+    LiveData<List<RealmDog>> dogs;
+    Observer<List<RealmDog>> observer = dogs -> {
+        frozenDogAdapter.updateData(dogs);
+    };
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        CustomApplication.getInjector(context).inject(this);
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_frozen, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        ButterKnife.bind(this, view);
+        frozenDogAdapter = new FrozenDogAdapter();
+        recyclerView.setHasFixedSize(true);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getActivity(), LinearLayoutManager.VERTICAL, false));
+        recyclerView.setAdapter(frozenDogAdapter);
+
+        dogs = monarchy.findAllFrozenWithChanges(realm -> realm.where(RealmDog.class));
+        dogs.observeForever(observer); // detach != destroy in fragments so this is manual
+    }
+
+    @Override
+    public void onDestroyView() {
+        dogs.removeObserver(observer);
+        super.onDestroyView();
+    }
+}